# adapter/in/rest/src/main/resources/openapi/comm-hub-api-v1.yaml
openapi: 3.1.0
...
info:
  title: Notification Hub — Source System API
  version: "1.0.0"
  description: |
    Приём сообщений и батчей от систем-источников Банка: единичные сообщения (в т.ч. OTP),
    батчевые рассылки со списком получателей, управление батчами и опрос статусов.

    Тело `POST /messages` — канонический входящий контракт IK-03 (§8.1), тот же самый документ,
    который системы-источники кладут в Kafka. Ошибки — RFC 9457 `application/problem+json`
    с машиночитаемым кодом в поле `code` (IR-01).
  contact:
    name: Notification Hub team
servers:
  - url: https://{host}/api/v1
    description: Контур Банка
    variables:
      host:
        default: comm-hub.hamkorbank.uz
...
paths:
...
  /batches/{batchId}:
    get:
      tags: [batches]
      summary: Состояние и прогресс батча
      operationId: getBatch
      parameters:
        - $ref: "#/components/parameters/BatchId"
      responses:
        "200":
          description: Состояние батча
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchStatus"
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
...
components:
  parameters:
    BatchId:
      name: batchId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Problem:
      description: Ошибка в формате RFC 9457 с кодом IR-01
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem+json; ветвиться следует по `code`, а не по HTTP-статусу
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code:
          type: string
          enum:
            - VALIDATION_FAILED
            - DUPLICATE
            - STREAM_SUSPENDED
            - QUOTA_EXCEEDED
            - TEMPLATE_NOT_PUBLISHED
            - TEMPLATE_VARIABLE_MISSING
            - SUPPRESSED
            - OPT_OUT
            - QUIET_HOURS
            - FREQUENCY_CAPPED
            - PAN_DETECTED
            - NO_ROUTE_AVAILABLE
            - KILL_SWITCH
            - SEND_STOPPED
            - TTL_EXPIRED
            - PROVIDER_REJECTED
            - ATTEMPTS_EXHAUSTED
            - NOT_FOUND
            - CONFLICT
            - RATE_LIMITED
            - INTERNAL_ERROR
        field:
          type: string
          description: Поле входящего документа, вызвавшее отказ
        messageId:
          type: string
          description: Идентификатор сообщения, к которому относится код (в т.ч. оригинал дубликата)
      required: [title, status, code]
...
    BatchProgress:
      type: object
      properties:
        total: { type: integer }
        processed: { type: integer }
        sent: { type: integer }
        delivered: { type: integer }
        failed: { type: integer }
        completionPercent: { type: number }
...
    BatchStatus:
      type: object
      properties:
        batchId: { type: string, format: uuid }
        streamId: { type: string }
        channel: { type: string, enum: [SMS, EMAIL, PUSH] }
        status:
          type: string
          enum: [ACCEPTED, PROCESSING, PAUSED, STOPPED, COMPLETED]
        total: { type: integer }
        progress: { $ref: "#/components/schemas/BatchProgress" }
        createdAt: { type: string, format: date-time }
        costEstimate: { type: [string, "null"] }
      required: [batchId, streamId, channel, status, total, progress, createdAt]